// src/interfaces/company_profile.ts
export default interface company_profile {
  symbol: string | null;
  price: number | null;
  market_cap: number | null;
  name: string | null;
  exchange: string | null;
  image: string | null;
  website: string | undefined;
  description: string | null;
  address: string | null;
  city: string | null;
  state: string | null;
  ceo: string | null;
  phone: string | null;
  change_number: number | null;
  change_percentage: number | null;
  num_of_employees: number | null;
  industry: string | null;
  sector: string | null;
}